Add webpack loader #114
